requirements_working.txt
absl-py==2.4.0
attrs==25.4.0
cffi==2.0.0
contourpy==1.3.3
cycler==0.12.1
filelock==3.25.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.15
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
packaging==26.0
pillow==12.1.1
protobuf==4.25.8
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sounddevice==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.10.0
typing_extensions==4.15.0
